create champ

+++ backend/src/main/java/com/greendrop/repository/CultureRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface CultureRepository extends JpaRepository<Culture, Integer> {
+public interface CultureRepository extends JpaRepository<Culture, Long> {
 }
